Refactor - move (internal) classes into namespaces
