+++ Toxon.UrlTemplates/Model/ExpressionEscapeMode.cs
@@ -1,0 +1,8 @@
+﻿namespace Toxon.UrlTemplates.Model
+{
+    internal enum ExpressionEscapeMode
+    {
+        Unreserved,
+        UnreservedAndReserved
+    }
+}
