+++ Toxon.UrlTemplates/Model/ExpressionVariableModifierExtensions.cs
@@ -1,35 +1,8 @@
 using System.Linq;
+using Toxon.UrlTemplates.Utils;
 
-namespace Toxon.UrlTemplates
+namespace Toxon.UrlTemplates.Model
 {
-    internal class ExpressionVariableModifier
-    {
-        private ExpressionVariableModifier() { }
-
-        public sealed class Prefix : ExpressionVariableModifier
-        {
-            public int Length { get; }
-
-            public Prefix(int length)
-            {
-                Length = length;
-            }
-
-            public override string ToString()
-            {
-                return $":{Length}";
-            }
-        }
-
-        public sealed class Explode : ExpressionVariableModifier
-        {
-            public override string ToString()
-            {
-                return "*";
-            }
-        }
-    }
-
     internal static class ExpressionVariableModifierExtensions
     {
         public static Option<int> GetPrefixLength(this ExpressionVariable variable)
