+++ Toxon.UrlTemplates/Model/LiteralComponent.cs
@@ -1,4 +1,4 @@
-﻿namespace Toxon.UrlTemplates
+﻿namespace Toxon.UrlTemplates.Model
 {
     internal class LiteralComponent : UrlTemplateComponent
     {
