+++ Toxon.UrlTemplates/Resolvers/ConstantDictionaryValue.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
+using Toxon.UrlTemplates.Values;
 
-namespace Toxon.UrlTemplates.Values
+namespace Toxon.UrlTemplates.Resolvers
 {
     public class ConstantDictionaryValue : IDictionaryValue
     {
