+++ Toxon.UrlTemplates/Toxon.UrlTemplates.csproj
@@ -40,25 +40,28 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="EnumerableExtensions.cs" />
-    <Compile Include="ExecutorState.cs" />
-    <Compile Include="ExpressionComponent.cs" />
-    <Compile Include="ExpressionOperator.cs" />
-    <Compile Include="ExpressionVariable.cs" />
-    <Compile Include="ExpressionVariableModifier.cs" />
+    <Compile Include="Execution\ExecutorState.cs" />
+    <Compile Include="Model\ExpressionComponent.cs" />
+    <Compile Include="Model\ExpressionEscapeMode.cs" />
+    <Compile Include="Model\ExpressionOperator.cs" />
+    <Compile Include="Model\ExpressionVariable.cs" />
+    <Compile Include="Model\ExpressionVariableModifier.cs" />
     <Compile Include="IValueResolver.cs" />
-    <Compile Include="LiteralComponent.cs" />
-    <Compile Include="Option.cs" />
-    <Compile Include="Parser.cs" />
-    <Compile Include="ParserError.cs" />
-    <Compile Include="ParserResult.cs" />
-    <Compile Include="ParserState.cs" />
-    <Compile Include="ParserUtils.cs" />
+    <Compile Include="Model\LiteralComponent.cs" />
+    <Compile Include="Model\ExpressionVariableModifierExtensions.cs" />
+    <Compile Include="Utils\Option.cs" />
+    <Compile Include="Parsing\Parser.cs" />
+    <Compile Include="Parsing\ParserError.cs" />
+    <Compile Include="Parsing\ParserResult.cs" />
+    <Compile Include="Parsing\ParserState.cs" />
+    <Compile Include="Parsing\ParserUtils.cs" />
+    <Compile Include="Parsing\ParserStateExtensions.cs" />
+    <Compile Include="Parsing\UrlChar.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UrlTemplate.cs" />
-    <Compile Include="UrlTemplateComponent.cs" />
-    <Compile Include="Executor.cs" />
-    <Compile Include="UrlTemplateOptimiser.cs" />
+    <Compile Include="Model\UrlTemplateComponent.cs" />
+    <Compile Include="Execution\Executor.cs" />
+    <Compile Include="Execution\ComponentOptimiser.cs" />
     <Compile Include="Values\ConstantArrayValue.cs" />
     <Compile Include="Values\ConstantDictionaryValue.cs" />
     <Compile Include="Values\ConstantStringValue.cs" />
@@ -72,6 +75,7 @@
     <None Include="paket.references" />
     <None Include="paket.template" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
