+++ Toxon.UrlTemplates/UrlTemplate.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Toxon.UrlTemplates.Execution;
+using Toxon.UrlTemplates.Model;
+using Toxon.UrlTemplates.Parsing;
 
 namespace Toxon.UrlTemplates
 {
